// lib/types.ts
export type PGSubmission = {
  createdat: string;
  dateofappt: string;
  firstlastname: string;
  email: string;
  phone: string;
  dob: string;
  tattoolocation: string;
  signaturedate: string;
  waiverdownloadurl: string;
};

export type Submission = {
  createdAt: string;
  dateOfAppt: string;
  firstLastName: string;
  email: string;
  phone: string;
  dob: string;
  tattooLocation: string;
  signatureDate: string;
  waiverDownloadUrl: string;
};